Introduce TypeTranslator
